identify double handler invocation's root cause

+++ src/Services/Basket/Basket.Api/Models/ShoppingCart.cs
@@ -7,7 +7,7 @@ namespace Basket.Api.Models
         [Identity]
         public string Username { get; set; } = default!;
         public List<ShoppingCartItem> Items { get; set; } = new();
-        public decimal TotalPrice => Items.Sum(x => x.Price * x.Quantity);
+        public int TotalPrice => Items.Sum(x => x.Price * x.Quantity);
 
         public ShoppingCart()
         {
