+++ src/Services/Discount/Discount.Grpc/Models/Coupon.cs
@@ -6,6 +6,6 @@ namespace Discount.Grpc.Models
         public int Id { get; set; }
         public string ProductName { get; set; } = default!;
         public string Description { get; set; } = default!;
-        public float Amount { get; set; }
+        public int Amount { get; set; }
     }
 }
